+++ include/DS_Joysticks.h
@@ -37,13 +37,13 @@ extern int DS_GetJoystickNumAxes (int joystick);
 extern int DS_GetJoystickNumButtons (int joystick);
 
 extern int DS_GetJoystickHat (int joystick, int hat);
-extern double DS_GetJoystickAxis (int joystick, int axis);
+extern float DS_GetJoystickAxis (int joystick, int axis);
 extern int DS_GetJoystickButton (int joystick, int button);
 
 extern void DS_JoysticksReset();
 extern void DS_JoysticksAdd (const int axes, const int hats, const int buttons);
 extern void DS_SetJoystickHat (int joystick, int hat, int angle);
-extern void DS_SetJoystickAxis (int joystick, int axis, double value);
+extern void DS_SetJoystickAxis (int joystick, int axis, float value);
 extern void DS_SetJoystickButton (int joystick, int button, int pressed);
 
 #ifdef __cplusplus
